Introduce IMetricsProvider
