+++ test/Elastic.Apm.Tests/MetricsTests.cs
@@ -1,5 +1,7 @@
-using System.Threading.Tasks;
+using System.Linq;
+using Elastic.Apm.Api;
 using Elastic.Apm.Metrics;
+using Elastic.Apm.Metrics.MetricsProvider;
 using Elastic.Apm.Tests.Mocks;
 using FluentAssertions;
 using Xunit;
@@ -9,7 +11,7 @@ namespace Elastic.Apm.Tests
 	public class MetricsTests
 	{
 		[Fact]
-		public async Task TestCollectAllMetrics()
+		public void CollectAllMetrics()
 		{
 			var mockPayloadSender = new MockPayloadSender();
 			var testLogger = new TestLogger();
@@ -18,6 +20,41 @@ namespace Elastic.Apm.Tests
 			mc.CollectAllMetrics();
 
 			mockPayloadSender.Metrics.Should().NotBeEmpty();
+		}
+
+		[Fact]
+		public void SystemCpu()
+		{
+			var testLogger = new TestLogger();
+
+			var systemTotalCpuProvider = new SystemTotalCpuProvider(testLogger);
+
+			//Needs to be called at least 2. times to deliver value - this is by design
+			systemTotalCpuProvider.GetValue();
+			var retVal = systemTotalCpuProvider.GetValue();
+			retVal.First().KeyValue.Value.Should().BeInRange(0, 1);
+		}
+
+		[Fact]
+		public void ProcessCpu()
+		{
+			var processTotalCpuProvider = new ProcessTotalCpuTimeProvider();
+
+			//Needs to be called at least 2. times to deliver value - this is by design
+			processTotalCpuProvider.GetValue();
+			var retVal = processTotalCpuProvider.GetValue();
+			retVal.First().KeyValue.Value.Should().BeInRange(0, 1);
+		}
+
+		[Fact]
+		public void GetWorkingSetAndVirtualMemory()
+		{
+			var processWorkingSetAndVirtualMemoryProvider = new ProcessWorkingSetAndVirtualMemoryProvider();
+			var retVal = processWorkingSetAndVirtualMemoryProvider.GetValue();
+
+			var enumerable = retVal as Sample[] ?? retVal.ToArray();
+			enumerable.Should().NotBeEmpty();
+			enumerable.First().KeyValue.Value.Should().BeGreaterThan(0);
 		}
 	}
 }
